phase4/requirements.txt
# Phase 4: LLM Recommendation and Reasoning Layer Dependencies
# Core LLM integration
groq>=0.5.0

# Data handling and validation
pydantic>=2.0.0

# Logging and utilities
python-dotenv>=1.0.0

# Optional: For enhanced JSON handling
orjson>=3.8.0

# Optional: For async support
aiohttp>=3.8.0

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
